[BUGFIX] Explain security mechanisms of the current property mapper
With the introduction of the "new" property mapper in
TYPO3 4.7 and the removal of the old property mapper in
TYPO3 6.2, the security mechanisms of forms have changed.

Therefore the documentation must no longer explain the
no longer existing `__hmac` field but the `__trustedProperties`
field instead.

Also, the @dontverifyrequesthash annotation is not worth
mentioning any more as it vanished with the old property
mapper as well.