Instantiation of the class Typo3QuerySettings (#482)
* Instantiation of the class Typo3QuerySettings

The Instantiation of the class Typo3QuerySettings has been changed because it comes with two mandatory parameters. If it is not called via GeneralUtility:makeinstance, the DI won't get triggered and as a result TYPO3 does not require the parameters. This leads to an exception regarding the missing parameters

* GeneralUtility namespace

GeneralUtility use namespace added

* Update Documentation/6-Persistence/2a-creating-the-repositories.rst

Co-authored-by: Mathias Brodala <mbrodala@pagemachine.de>