Remove warning
Remove warning on the page "Individual database query". This warning
comes right in the middle of explaing the order and Extbase execute()
and uses Doctrine DBAL persistence without really explaining why.

This made the already lengthy page confusing without really making
it clear why this warning exists.

If there is the need to use QueryBuilder database persistence
functionality it is probably best to introduce this and handle it
on a separate page.

Resolves: #240