Add notice with links to newer Extbase / Fluid content
Also, the Includes.rst.txt was not included on some
pages. This was added as well.

The manual is currently only rendered up to 10.4. There
is no version selector to switch to a newer version and
there is not easy way to get to the newer version of a
page unless one starts another search.

While a link on each page to link to newer content of this
page would be preferable (from point of view of the reader),
this is not feasible to do for every page.
However stating where the new content should be ("TYPO3
Explained") might be helpful when searching for it and
adding a link to the respective chapters might be
sufficient.

(Unfortunately, when googling one often ends up on older
pages unless this is specifically prevented by specifying
the versions).

In addition to this, a link to the newer content of the page
might be helpful on some of the pages. Here, it might be a
good idea to check which pages are viewed most frequently.

Resolves: #590